requirements.txt
# Diabetic Retinopathy Detection System
# Python 3.7+

# Deep Learning
tensorflow>=2.0.0
keras>=2.4.0

# Image Processing
opencv-python>=4.0.0
Pillow>=8.0.0
scikit-image>=0.18.0

# Data Manipulation
numpy>=1.19.0
pandas>=1.2.0

# Machine Learning Utilities
scikit-learn>=0.24.0

# Web Application
flask>=2.0.0
flask-cors>=3.0.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# PDF Generation
reportlab>=3.5.0
PyPDF2>=1.26.0

# Utilities
tqdm>=4.50.0
Pillow>=8.0.0

# Optional: GPU Support
# tensorflow-gpu>=2.0.0
